+++ renderer/units/Unihan Data Finder/grid-data-table.js
@@ -1,6 +1,7 @@
 //
 const regexp = require ('../../lib/unicode/regexp.js');
 const unicode = require ('../../lib/unicode/unicode.js');
+const unihanData = require ('../../lib/unicode/parsed-unihan-data.js');
 //
 const deferredSymbols = false;
 //
@@ -102,11 +103,28 @@ module.exports.create = function (characters, params, highlightedCharacter)
             {
                 let isUnified = regexp.isUnified (character);
                 let status = isUnified ? "Unified Ideograph" : "Compatibility Ideograph";
+                let set = "Full Unihan";
+                let tags = unihanData.codePoints[data.codePoint];
+                if ("kIICore" in tags)
+                {
+                    set = "IICore";
+                }
+                else if ("kUnihanCore2020" in tags)
+                {
+                    set = "Unihan Core (2020)";
+                }
                 if (isUnified && /[\uF900-\uFAFF]/.test (character))
                 {
                     symbolCell.classList.add ('misclassified');
                 }
-                symbolCell.title = `Code Point: ${data.codePoint}\nAge: Unicode ${data.age} (${data.ageDate})\nStatus: ${status}`;
+                let lines =
+                [
+                    `Code Point: ${data.codePoint}`,
+                    `Age: Unicode ${data.age} (${data.ageDate})`,
+                    `Set: ${set}`,
+                    `Status: ${status}`
+                ];
+                symbolCell.title = lines.join ("\n");
                 if (deferredSymbols)
                 {
                     symbol.textContent = "\u3000";  // Ideographic space